:truck: - chore: moved utils.py to common module
