+++ project_name/apps/core/views.py
@@ -8,7 +8,7 @@ from rest_framework.response import Response
 from project_name.common.serializers import ErrorSerializer
 from project_name.apps.core.handlers.list_locations import list_locations
 from project_name.apps.core.serializers import LocationSerializer
-from project_name.utils import read_docs_md
+from project_name.common.utils import read_docs_md
 
 logger = logging.getLogger('watchtower-logger')
